pyproject.toml
[project]
name = "ai-ml-forex-predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README_TECHNICAL.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "scikit-learn>=1.8.0",
    "torch>=2.11.0",
    "uvicorn>=0.45.0",
    "xgboost>=3.2.0",
    "yfinance>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
]
